requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
flask
flask-cors
pymongo
dnspython
python-dotenv